MVVM Phase 1
