+++ apps/frontend/src/__tests__/layout-providers.test.tsx
@@ -56,6 +56,7 @@ vi.mock('@semiont/react-ui', async () => {
     }),
     kbBackendUrl: (kb: any) => `${kb.protocol}://${kb.host}:${kb.port}`,
     getKbSessionStatus: () => 'authenticated',
+    useJobReplayBridge: () => {},
   };
 });
