+++ packages/react-ui/src/hooks/__tests__/useHoverDelay.test.ts
@@ -11,7 +11,7 @@
 import { describe, it, expect, beforeEach, afterEach } from 'vitest';
 import { renderHook, act } from '@testing-library/react';
 import { useHoverDelay } from '../useHoverDelay';
-import { HOVER_DELAY_MS } from '../useBeckonFlow';
+import { HOVER_DELAY_MS } from '../useHoverEmitter';
 
 describe('useHoverDelay', () => {
   beforeEach(() => {
