+++ packages/react-ui/src/index.ts
@@ -39,6 +39,7 @@ export * from './hooks/useObservableBrowse';
 export * from './hooks/usePanelWidth';
 export * from './hooks/useResourceEvents';
 export * from './hooks/useGlobalEvents';
+export * from './hooks/useJobReplayBridge';
 export * from './hooks/useRovingTabIndex';
 export * from './hooks/useSessionExpiry';
 export * from './contexts/ThemeContext';
@@ -216,7 +217,9 @@ export * from './features/resource-compose/components/ResourceComposePage';
 export * from './features/resource-discovery/components/ResourceDiscoveryPage';
 export * from './features/resource-discovery/components/ResourceCard';
 export * from './features/resource-viewer/components/ResourceViewerPage';
-export * from './hooks/useBeckonFlow';
+export * from './hooks/useHoverEmitter';
+export * from './hooks/useAttentionStream';
+export { createBeckonVM, type BeckonVM, createHoverHandlers, type HoverHandlers, HOVER_DELAY_MS } from '@semiont/api-client';
 export * from './hooks/useMarkFlow';
 export * from './hooks/usePanelBrowse';
 export * from './hooks/useYieldFlow';
